updata 2010/10/27 02:53.

+++ net/ptree.h
@@ -21,7 +21,7 @@ struct ptree_node {
   struct  ptree_mask *rn_mklist;
   caddr_t key;
   caddr_t rn_mask;
-  int   keylen;
+  short   keylen;
 
   short	rn_bit;
   char	rn_bmask;
